restructed directories
